refactor(server): inline Identity.verify impl (not a handler)
verify() mounts no routes, so it doesn't belong under internal/handlers.
Inline the implementation directly in Identity.ts, mirroring Kv.ts and
CliAuth.ts; Handler.auth imports it from ../../Identity.js.

Amp-Thread-ID: https://ampcode.com/threads/T-019ea89f-ae30-73df-925c-ca505c59997f
